Add: Authorization with Policies

+++ app/Http/Controllers/Api/EventController.php
@@ -27,7 +27,7 @@ class EventController extends Controller
      */
     public function index()
     {
-        // Gate::authorize('viewAny', Event::class);
+        Gate::authorize('viewAny', Event::class);
         $query = $this->loadRelationships(Event::query());
 
         return EventResource::collection(
@@ -40,7 +40,7 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
-        // Gate::authorize('create', Event::class);
+        Gate::authorize('create', Event::class);
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
@@ -59,7 +59,7 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
-        // Gate::authorize('view', $event);
+        Gate::authorize('view', $event);
         $event->load('user', 'attendees');
         return new EventResource(
             $this->loadRelationships($event)
@@ -71,7 +71,7 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
-        Gate::authorize('update-event', $event);
+        Gate::authorize('update', $event);
 
         // if (Gate::denies('update-event', $event)) {
         //     abort(403, 'You are not authorized to update this event.');
@@ -94,7 +94,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
-        // Gate::authorize('delete-event', $event);
+        Gate::authorize('delete', $event);
         $event->delete();
 
         return response(status: 204);
